+++ kroto-plus-protoc/src/main/kotlin/com/github/marcoferrer/krotoplus/KrotoPlusProtoCMain.kt
@@ -1,18 +1,20 @@
 @file:JvmName("KrotoPlusProtoCMain")
 package com.github.marcoferrer.krotoplus
 
-import com.github.marcoferrer.krotoplus.generators.Generator
-import com.github.marcoferrer.krotoplus.generators.GrpcStubExtsGenerator
-import com.github.marcoferrer.krotoplus.generators.ProtoBuildersGenerator
+import com.github.marcoferrer.krotoplus.generators.*
 import com.google.protobuf.compiler.PluginProtos
+import java.io.File
 
 object Manifest {
-    val name        = javaClass.`package`.name.orEmpty()
-    val implTitle   = javaClass.`package`.implementationTitle.orEmpty()
-    val implVersion = javaClass.`package`.implementationVersion.orEmpty()
+    val name        = this::class.java.`package`.name.orEmpty()
+    val implTitle   = this::class.java.`package`.implementationTitle.orEmpty()
+    val implVersion = this::class.java.`package`.implementationVersion.orEmpty()
 }
 
 fun main(args: Array<String>) {
+
+//    val file = File("/Users/mferrer/IdeaProjects/kroto-plus/kroto-plus-proto-c/sample_plugin_request.bin")
+//    val protoRequest = PluginProtos.CodeGeneratorRequest.parseFrom(file.inputStream())
 
     val protoRequest = PluginProtos.CodeGeneratorRequest.parseFrom(System.`in`)
 
@@ -22,7 +24,9 @@ fun main(args: Array<String>) {
 
     val generators = listOf(
             GrpcStubExtsGenerator(context),
-            ProtoBuildersGenerator(context)
+            ProtoBuildersGenerator(context),
+            ExtendableMessagesGenerator(context),
+            MockServiceGenerator(context)
     )
 
     generators