+++ kroto-plus-protoc/src/main/kotlin/com/github/marcoferrer/krotoplus/schema/DescriptorProtoUtils.kt
@@ -3,23 +3,15 @@ package com.github.marcoferrer.krotoplus.schema
 import com.github.marcoferrer.krotoplus.Manifest
 import com.google.common.base.CaseFormat
 import com.google.protobuf.DescriptorProtos
+import com.google.protobuf.compiler.PluginProtos
 import com.squareup.kotlinpoet.AnnotationSpec
-import com.squareup.kotlinpoet.asClassName
+import com.squareup.kotlinpoet.ClassName
 
-//fun DescriptorProtos.DescriptorProto.canonicalProtoName(protoFile: DescriptorProtos.FileDescriptorProto) =
-//        if(protoFile.hasPackage())
-//            "${protoFile.`package`}.${this.name}" else this.name
-//
-//fun DescriptorProtos.DescriptorProto.canonicalJavaClassName(protoFile: DescriptorProtos.FileDescriptorProto): String{
-//
-//    val pkg = protoFile.javaPackage.let {
-//        if(it.isEmpty()) "" else "$it."
-//    }
-//
-//    return if(protoFile.options.javaMultipleFiles)
-//        "$pkg${this.name}" else
-//        "$pkg${protoFile.javaOuterClassname}.${this.name}"
-//}
+val DescriptorProtos.MethodDescriptorProto.isEmptyInput
+    get() = this.inputType == ".google.protobuf.Empty"
+
+val DescriptorProtos.MethodDescriptorProto.isNotEmptyInput
+    get() = !isEmptyInput
 
 fun DescriptorProtos.DescriptorProto.outputPath(protoFile: DescriptorProtos.FileDescriptorProto): String {
 
@@ -45,7 +37,6 @@ val DescriptorProtos.FileDescriptorProto.javaOuterClassname: String
                     "${outerClassname}OuterClass" else outerClassname
             }
 
-
 val DescriptorProtos.FileDescriptorProto.javaPackage: String
     get() = when{
         this.options.hasJavaPackage() -> this.options.javaPackage
@@ -56,8 +47,27 @@ val DescriptorProtos.FileDescriptorProto.javaPackage: String
 val DescriptorProtos.FileDescriptorProto.protoPackage: String
     get() = if (this.hasPackage()) this.`package` else ""
 
+fun PluginProtos.CodeGeneratorRequest.getAllProtoTypes(): Sequence<ProtoType> =
+        protoFileList
+                .asSequence()
+                .flatMap { fileDescriptor ->
+                    val protoMessages = fileDescriptor
+                            .messageTypeList
+                            .asSequence()
+                            .map { ProtoMessage(it,fileDescriptor) }
+                            .flattenProtoTypes()
+
+                    val protoEnums = fileDescriptor
+                            .enumTypeList
+                            .asSequence()
+                            .map { ProtoEnum(it,fileDescriptor) }
+                            .flattenProtoTypes()
+
+                    protoMessages + protoEnums
+                }
+
 fun DescriptorProtos.FileDescriptorProto.getGeneratedAnnotationSpec() =
-        AnnotationSpec.builder(javax.annotation.Generated::class.asClassName())
+        AnnotationSpec.builder(ClassName("javax.annotation","Generated"))
                 .addMember("value = [%S]", "by ${Manifest.implTitle} (version ${Manifest.implVersion})")
                 .addMember("comments = %S", "Source: $name")
                 .build()